Flask_Backend/requirements.txt
flask
python-dotenv
streamlit
google-generativeai
langchain
PyPDF2
chromadb